Extract serve discovery into a shared module
Move the .pyric/serve.json pointer + identity-pinned health probing out
of cli/mcp-proxy.ts into serve/discovery.ts so the Node remote-sandbox
client shares the exact same discovery rules (both loopback families,
instanceId pinning against cross-family port squatters, legible
failure) instead of a drifting copy. mcp-proxy re-exports discoverServe
for existing consumers; behavior unchanged.
