+++ packages/pyric-tools/src/cli/mcp-proxy.ts
@@ -36,18 +36,14 @@
  * window as a fallback. Degrades LEGIBLY: if no serve is found, or the pointed
  * server's identity can't be matched, we report it — never a silent hang.
  */
-import { existsSync, readFileSync } from 'node:fs';
-import { join } from 'node:path';
 import type { JSONRPCMessage } from '@modelcontextprotocol/sdk/types.js';
 import type { ParsedArgs } from './parse-args.js';
+import { discoverServe, SCAN_PORTS } from '../serve/discovery.js';
 
-/** Ports probed when the pointer is absent — serve's default scan window PLUS
- *  the vite-plugin default (5174), which the old 5000-5004 window missed (so a
- *  vite-only project was never found by scan). The standalone `pyric bridge`
- *  serves `/health` (not `/__pyric/health`) and writes no pointer, so it is
- *  registered directly via `claude mcp add`, not discovered here. */
-const SCAN_PORTS = [5000, 5001, 5002, 5003, 5004, 5174];
-const POINTER = join('.pyric', 'serve.json');
+// Discovery (pointer + identity-pinned health probing) lives in
+// `serve/discovery.ts` — shared with the Node remote-sandbox client
+// (`remote/index.ts`). Re-exported here for existing consumers.
+export { discoverServe } from '../serve/discovery.js';
 
 /** A relayed request with no response within this window is failed with a
  *  JSON-RPC error rather than hanging forever. Sits just above the bridge's
@@ -55,142 +51,6 @@ const POINTER = join('.pyric', 'serve.json');
 const REQUEST_TIMEOUT_MS = 35_000;
 /** JSON-RPC error code for proxy-synthesized failures (server-defined range). */
 const PROXY_ERROR_CODE = -32001;
-
-interface HealthLite {
-  mode?: string;
-  instanceId?: string;
-}
-interface Discovered {
-  mcpUrl: string;
-  url: string;
-  /** `http://<family>:<port>` the server actually answered on. */
-  base: string;
-  /** Identity pinned at discovery. Null only when talking to an older server
-   *  that predates the instanceId field (matching is then skipped). */
-  instanceId: string | null;
-  source: string;
-}
-
-/**
- * Probe BOTH loopback families on a port and return the base that answers.
- *
- * Hostname-based URLs are a trap here: serve writes `http://localhost:...`
- * for humans (browsers dual-stack fine), but `localhost` resolution differs
- * by runtime — node/undici prefers IPv6 `::1`, and a serve under one runtime
- * may bind `127.0.0.1`-only while under another binds `::1`-only. So the
- * proxy never trusts the hostname: it takes the PORT and tries explicit
- * `127.0.0.1` and `[::1]`, using whichever the server is actually on.
- */
-function basesForPort(port: number): string[] {
-  return [`http://127.0.0.1:${port}`, `http://[::1]:${port}`];
-}
-
-async function probeHealth(base: string): Promise<HealthLite | null> {
-  try {
-    const res = await fetch(`${base}/__pyric/health`, { signal: AbortSignal.timeout(1000) });
-    if (res.status !== 200) return null;
-    const body = (await res.json()) as HealthLite;
-    return body.mode === 'sandbox' ? body : null;
-  } catch {
-    return null;
-  }
-}
-
-/**
- * First loopback base on `port` whose health reports a sandbox bridge. With an
- * `expectedInstanceId`, returns ONLY a family whose health identity matches —
- * so when two sandboxes collide on one port across families, the proxy locks
- * onto the one the pointer names, not merely the first to answer. (An older
- * server with no `instanceId` field can't be identity-checked; matching is
- * skipped in that case so the pointer still resolves.)
- */
-async function healthyBase(
-  port: number,
-  expectedInstanceId?: string | null,
-): Promise<{ base: string; instanceId: string | null } | null> {
-  for (const base of basesForPort(port)) {
-    const health = await probeHealth(base);
-    if (!health) continue;
-    const id = health.instanceId ?? null;
-    // Healthy but the WRONG server (the cross-family squatter): skip, try next.
-    if (expectedInstanceId && id !== expectedInstanceId) continue;
-    return { base, instanceId: id };
-  }
-  return null;
-}
-
-/** Best-effort port extraction from a pointer url/mcpUrl. */
-function portOf(u: string | undefined): number | null {
-  const m = u?.match(/:(\d{2,5})(?:\/|$)/);
-  return m ? Number(m[1]) : null;
-}
-
-/** Find the running serve: pointer first (in `cwd`), then a port scan. The
- *  pointer gives the PORT and (when present) the identity; the family is
- *  resolved by probing, so the returned base always uses the address the
- *  server is actually reachable on. */
-export async function discoverServe(
-  cwd: string,
-  log: (m: string) => void = () => {},
-  // Injectable so discovery can be tested hermetically — the default scan probes
-  // real localhost ports, which a test environment can't guarantee are free.
-  scanPorts: number[] = SCAN_PORTS,
-): Promise<Discovered | null> {
-  const pointerPath = join(cwd, POINTER);
-  if (existsSync(pointerPath)) {
-    try {
-      const p = JSON.parse(readFileSync(pointerPath, 'utf8')) as {
-        url?: string;
-        mcpUrl?: string;
-        port?: number;
-        instanceId?: string;
-      };
-      const port = p.port ?? portOf(p.mcpUrl) ?? portOf(p.url);
-      if (port) {
-        const expectedId = typeof p.instanceId === 'string' && p.instanceId ? p.instanceId : null;
-        const hit = await healthyBase(port, expectedId);
-        if (hit) {
-          return {
-            mcpUrl: `${hit.base}/__pyric/mcp`,
-            url: hit.base,
-            base: hit.base,
-            instanceId: hit.instanceId,
-            source: `pointer ${POINTER}`,
-          };
-        }
-        // The pointer named a specific identity we could NOT find on its port:
-        // a different sandbox may be squatting it (cross-family collision) or
-        // the server stopped. Do NOT scan into a possibly-wrong server — that
-        // split-brain is exactly what this identity check prevents. Fail legibly.
-        if (expectedId) {
-          log(
-            `pointer ${POINTER} names a server (instanceId ${expectedId.slice(0, 8)}…) ` +
-              `that isn't answering on port ${port} — another sandbox may be squatting the ` +
-              `port on the other loopback family, or the server stopped. Not falling back to ` +
-              `a blind port scan (it could hit the wrong sandbox). Restart your dev server, ` +
-              `and open http://127.0.0.1:${port} (NOT localhost) so a single family is used.`,
-          );
-          return null;
-        }
-      }
-    } catch {
-      /* stale/corrupt pointer — fall through to scan */
-    }
-  }
-  for (const port of scanPorts) {
-    const hit = await healthyBase(port);
-    if (hit) {
-      return {
-        mcpUrl: `${hit.base}/__pyric/mcp`,
-        url: hit.base,
-        base: hit.base,
-        instanceId: hit.instanceId,
-        source: `port scan (:${port})`,
-      };
-    }
-  }
-  return null;
-}
 
 // ── JSON-RPC shape helpers (relay-level; no SDK runtime import) ──────────────
 function msgId(m: JSONRPCMessage): string | number | null {
